requirements.txt
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
lxml==4.5.0
mongoengine==0.19.1
mysqlclient==1.4.6
Pillow==7.1.1
PyJWT==1.7.1
pymongo==3.10.1
python-docx==0.8.10
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
xlrd==1.2.0
xlwt==1.3.0
